Adding the Sycamore design files to git
These include lots of notes and resources for
bits and pieces that make up Sycamore, along
with a changelog of sorts (More of a bug list...).

The design files here have the original Mutable
Instruments design embedded, with the PCBs for the
front panel and main PCB in a single kicad_pcb
file.