Adding Rowan's firmware for the 1.0 hardware
This isn't perfect, but seems to do the job. Rowan can:
-React to events over i2c
-Update asynchronously, after i2c messages are receieved

There are definitely races and bugs here in the interrupt-based i2c handling, and ideally there needs to be a FIFO of operations to perform rather than a single i2c value. However, experimentally with Sycamore it seems like the CircuitPython code is slow enough than Rowan is quite capable of processing the whole interrupt very quickly.